+++ src/i-call-service.extension.ts
@@ -1,106 +1,19 @@
 import { DOWN, is, TServiceParams, UP } from "@digital-alchemy/core";
-import {
-  HassServiceDTO,
-  ServiceListField,
-  ServiceListFieldDescription,
-  ServiceListServiceTarget,
-} from "@digital-alchemy/hass";
-import { factory, SyntaxKind, TypeElement, TypeNode } from "typescript";
+import { HassServiceDTO, ServiceListField, ServiceListServiceTarget } from "@digital-alchemy/hass";
+import { factory, SyntaxKind, TypeElement } from "typescript";
 
 export async function ICallServiceExtension({ hass, type_writer }: TServiceParams) {
   return async function () {
     const domains = await hass.fetch.listServices();
 
-    // #MARK: fieldPropertySignature
-    function fieldPropertySignature(
-      parameterName: string,
-      { selector, ...details }: ServiceListFieldDescription,
-      serviceDomain: string,
-      serviceName: string,
-    ) {
-      let node: TypeNode;
-      const { domain } = selector?.entity ?? {};
-      // : boolean
-      if (!is.undefined(selector?.boolean))
-        node = factory.createKeywordTypeNode(SyntaxKind.BooleanKeyword);
-      // : number
-      else if (!is.undefined(selector?.number))
-        node = factory.createKeywordTypeNode(SyntaxKind.NumberKeyword);
-      // : string
-      else if (!is.undefined(selector?.text) || !is.undefined(selector?.time))
-        node = factory.createKeywordTypeNode(SyntaxKind.StringKeyword);
-      // string | `domain.${keyof typeof ENTITY_SETUP.domain}`
-      else if (!is.undefined(selector?.entity))
-        // some combination of: PICK_ENTITY | PICK_ENTITY[] | PICK_ENTITY<"domain"> | PICK_ENTITY<"domain">[]
-        node = type_writer.entity.buildEntityReference(domain, selector);
-      // : "option" | "option" | "option" | "option"
-      else if (!is.undefined(selector?.select))
-        node = factory.createUnionTypeNode(
-          selector?.select.options.map((i: string | Record<"label" | "value", string>) =>
-            factory.createLiteralTypeNode(factory.createStringLiteral(is.string(i) ? i : i.value)),
-          ),
-        );
-      // : Record<string, unknown> | (unknown[]);
-      else if (is.undefined(selector?.object)) {
-        node = factory.createKeywordTypeNode(SyntaxKind.UnknownKeyword);
-      }
-      // else if (!is.undefined(selector?.))
-      // : unknown
-      else {
-        node = handleSelectors(serviceDomain, serviceName, {
-          selector,
-          ...details,
-        });
-      }
-
-      return type_writer.tsdoc.parameterComment(
-        factory.createPropertySignature(
-          undefined,
-          factory.createIdentifier(parameterName),
-          details.required ? undefined : factory.createToken(SyntaxKind.QuestionToken),
-          node,
-        ),
-        parameterName,
-        { selector, ...details },
-      );
-    }
-
-    // #MARK: handleSelectors
-    function handleSelectors(
-      serviceDomain: string,
-      serviceName: string,
-      { selector }: ServiceListFieldDescription,
-    ) {
-      if ("object" in selector && selector.object === null) {
-        // if (serviceDomain === "conversation") {
-        // console.log({ serviceDomain, serviceName }, selector, options);
-        // }
-        return factory.createKeywordTypeNode(SyntaxKind.UnknownKeyword);
-      }
-
-      return factory.createUnionTypeNode([
-        serviceDomain === "scene" && serviceName === "apply"
-          ? factory.createTypeReferenceNode(factory.createIdentifier("Partial"), [
-              factory.createTypeReferenceNode(factory.createIdentifier("Record"), [
-                factory.createTypeReferenceNode(factory.createIdentifier("PICK_ENTITY"), undefined),
-                factory.createKeywordTypeNode(SyntaxKind.UnknownKeyword),
-              ]),
-            ])
-          : factory.createTypeReferenceNode(factory.createIdentifier("Record"), [
-              factory.createKeywordTypeNode(SyntaxKind.StringKeyword),
-              factory.createKeywordTypeNode(SyntaxKind.UnknownKeyword),
-            ]),
-        factory.createParenthesizedType(
-          factory.createArrayTypeNode(factory.createKeywordTypeNode(SyntaxKind.UnknownKeyword)),
-        ),
-      ]);
-    }
-
     // #MARK: BuildServiceParameters
     function serviceParameters(domain: string, key: string, value: ServiceListField) {
       return [
-        // f( service_data: { ...definition } )
-        //    Provide this        ^^^^^^
+        // Build contents of object:
+        //
+        // f(service_data: {
+        //   [service]: [property signature]
+        //  })
         factory.createParameterDeclaration(
           undefined,
           undefined,
@@ -113,7 +26,9 @@ export async function ICallServiceExtension({ hass, type_writer }: TServiceParam
             [
               ...Object.entries(value.fields)
                 .sort(([a], [b]) => (a > b ? UP : DOWN))
-                .map(([service, details]) => fieldPropertySignature(service, details, domain, key)),
+                .map(([service, details]) =>
+                  type_writer.fields.fieldPropertySignature(service, details, domain, key),
+                ),
               type_writer.entity.createTarget(value.target as ServiceListServiceTarget),
             ].filter(i => !is.undefined(i)) as TypeElement[],
           ),
