+++ src/main.ts
@@ -4,6 +4,7 @@ import { LIB_HASS } from "@digital-alchemy/hass";
 
 import { BuildTypes } from "./build.extension";
 import { EntityReference } from "./entity-reference.extension";
+import { FieldBuilder } from "./field-builder.extension";
 import { ICallServiceExtension } from "./i-call-service.extension";
 import { Identifiers } from "./identifiers.extension";
 import { Printer } from "./printer.extension";
@@ -22,6 +23,7 @@ export const TYPE_WRITER = CreateApplication({
     build: BuildTypes,
     call_service: ICallServiceExtension,
     entity: EntityReference,
+    fields: FieldBuilder,
     identifiers: Identifiers,
     printer: Printer,
     tsdoc: TSDoc,
